% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/node.R
\name{find_node}
\alias{find_node}
\title{Find the \command{node} executable}
\usage{
find_node(cache = TRUE, dir = NULL, version = NULL)
}
\arguments{
\item{cache}{Whether to search for \command{node} again if a node
directory containing the \command{node} executable of the expected
version (if provided) has been found previously. Search again if
\code{cache = FALSE}.}

\item{dir}{A character vector of potential directory paths under which
\command{node} may be found.}

\item{version}{The version of node to look for (e.g., \code{"14.15.4"}). If
not provided, this function searches for the highest version under the
potential directories.}
}
\value{
A list containing the directory and version of node (if found).
}
\description{
Searches for the \command{node} executable in a few places and use the
highest version found, unless a specific version is requested.
}
\examples{
find_node()
find_node(version = '10.13.0')
}
