% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mathjax_node_setup.R
\name{mathjax_npm_root}
\alias{mathjax_npm_root}
\alias{mathjax_npm_available}
\alias{uninstall_mathjax_npm}
\alias{install_mathjax_npm}
\title{Install/Uninstall "mathjax-node"}
\usage{
mathjax_npm_root()

mathjax_npm_available()

uninstall_mathjax_npm()

install_mathjax_npm(force = FALSE)
}
\arguments{
\item{force}{Whether to force to install (override) "mathjax-node".}
}
\value{
\code{mathjax_npm_root} returns a single character value,
\code{mathjax_npm_available} returns a single logical value,
\code{uninstall_mathjax_npm()} and \code{install_mathjax_npm()} return
a \code{NULL} value.
}
\description{
The function \code{install_mathjax_npm()} downloads and installs "mathjax-node" ,
(a "JavaScript" display engine for mathematics) in the user data directory.

The function \code{uninstall_mathjax_npm()} removes "mathjax-node" from the user
data directory.

\code{mathjax_npm_root()} returns the root directory of "mathjax-node" in the user
data directory and \code{mathjax_npm_available()} checks if "mathjax-node" is
in the user data directory.
}
\examples{
if(node_available() && !mathjax_npm_available()){
  install_mathjax_npm()
  uninstall_mathjax_npm()
}
}
