% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/node.R
\name{node_available}
\alias{node_available}
\alias{node_version}
\title{Check node availability and version}
\usage{
node_available(version = NULL, error = FALSE)

node_version()
}
\arguments{
\item{version}{Required version of node}

\item{error}{Whether to signal an error if node with the required version
is not found}
}
\value{
\code{node_available} returns a logical indicating whether the
required version of node is available. \code{node_version} returns a
\code{\link[base]{numeric_version}} with the version of node found.
}
\description{
\code{node_available()} determine whether node is currently available on
the system (optionally checking for a specific version or greater).

\code{node_version()} determine the specific version of node available.
}
\examples{
\dontrun{
library(equatags)

if (node_available())
  cat("node", as.character(node_version()), "is available!\n")

if (node_available("10.13.0"))
  cat("required version of node is available!\n")
}
}
