% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mathjax_node_setup.R
\name{mathjax_available}
\alias{mathjax_available}
\title{Is 'mathjax-node' available}
\usage{
mathjax_available()
}
\value{
a single logical value.
}
\description{
Checks if 'mathjax-node' is available within a directory of the user.
}
\examples{
mathjax_available()
}
\seealso{
Other tools for 'mathjax-node': 
\code{\link{mathjax_install}()},
\code{\link{mathjax_uninstall}()}
}
\concept{tools for 'mathjax-node'}
