% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mathjax_node_setup.R
\name{mathjax_uninstall}
\alias{mathjax_uninstall}
\title{Uninstall 'mathjax-node'}
\usage{
mathjax_uninstall()
}
\value{
a single logical value, FALSE if the operation
failed, TRUE otherwise.
}
\description{
Removes 'mathjax-node'.
}
\examples{
library(locatexec)

if(exec_available("npm") &&
   mathjax_available()) {
  mathjax_uninstall()
  mathjax_install()
}
}
\seealso{
Other tools for 'mathjax-node': 
\code{\link{mathjax_available}()},
\code{\link{mathjax_install}()}

Other tools for 'mathjax-node': 
\code{\link{mathjax_available}()},
\code{\link{mathjax_install}()}
}
\concept{tools for 'mathjax-node'}
