% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_mathjax.R
\name{transform_mathjax}
\alias{transform_mathjax}
\title{'MathJax' equation as 'HTML' or 'MathML'.}
\usage{
transform_mathjax(x, to = "html")
}
\arguments{
\item{x}{MathJax equations}

\item{to}{output format, one of 'html' or 'mml'}
}
\value{
a character vector that contains 'html' or 'mml'
codes corresponding to the equations.
}
\description{
Get 'HTML' or 'MathML' XML codes corresponding to the rendering
of 'MathJax' equations.
}
\examples{
x <- c("(ax^2 + bx + c = 0)",
       "x = {-b \\\\pm \\\\sqrt{b^2-4ac} \\\\over 2a}.")
z <- transform_mathjax(x = x, to = "html")
cat(z, sep = "\n\n")
z <- transform_mathjax(x = x, to = "mml")
cat(z, sep = "\n\n")
}
